+++ Source/Modules/Catalog/Gapotchenko.FX/ExceptionPolyfills.cs
@@ -6,10 +6,6 @@
 // File introduced by: Oleksiy Gapotchenko
 // Year of introduction: 2025
 
-#if NET8_0_OR_GREATER
-#define TFF_EXCEPTION_THROWIF
-#endif
-
 using Gapotchenko.FX.Properties;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
@@ -26,6 +22,75 @@ namespace Gapotchenko.FX;
 public static class ExceptionPolyfills
 {
     /// <summary>
+    /// Provides polyfill extension methods for <see cref="ArgumentException"/> class.
+    /// </summary>
+    extension(ArgumentException)
+    {
+        /// <summary>
+        /// Throws an exception if <paramref name="argument"/> is <see langword="null"/> or empty.
+        /// </summary>
+        /// <param name="argument">The string argument to validate as non-null and non-empty.</param>
+        /// <param name="paramName">The name of the parameter with which <paramref name="argument"/> corresponds.</param>
+        /// <exception cref="ArgumentNullException"><paramref name="argument"/> is <see langword="null"/>.</exception>
+        /// <exception cref="ArgumentException"><paramref name="argument"/> is empty.</exception>
+#if NET7_0_OR_GREATER
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+#endif
+        public static void ThrowIfNullOrEmpty([NotNull] string? argument, [CallerArgumentExpression(nameof(argument))] string? paramName = null)
+#if !NETCOREAPP3_0_OR_GREATER
+#pragma warning disable CS8777 // Parameter must have a non-null value when exiting.
+#endif
+        {
+#if NET7_0_OR_GREATER
+            ArgumentException.ThrowIfNullOrEmpty(argument, paramName);
+#else
+            if (string.IsNullOrEmpty(argument))
+                Throw(argument, paramName);
+
+            [DoesNotReturn]
+            static void Throw(string? argument, string? paramName)
+            {
+                ArgumentNullException.ThrowIfNull(argument, paramName);
+                throw new ArgumentException(Resources.Argument_EmptyString, paramName);
+            }
+#endif
+        }
+#pragma warning restore CS8777 // Parameter must have a non-null value when exiting.
+
+        /// <summary>
+        /// Throws an exception if <paramref name="argument"/> is <see langword="null"/> or empty, or consists only of white-space characters.
+        /// </summary>
+        /// <param name="argument">The string argument to validate.</param>
+        /// <param name="paramName">The name of the parameter with which <paramref name="argument"/> corresponds.</param>
+        /// <exception cref="ArgumentNullException"><paramref name="argument"/> is <see langword="null"/>.</exception>
+        /// <exception cref="ArgumentException"><paramref name="argument"/> is empty or consists only of white-space characters.</exception>
+#if NET8_0_OR_GREATER
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+#endif
+        public static void ThrowIfNullOrWhiteSpace([NotNull] string? argument, [CallerArgumentExpression(nameof(argument))] string? paramName = null)
+#if !NETCOREAPP3_0_OR_GREATER
+#pragma warning disable CS8777 // Parameter must have a non-null value when exiting.
+#endif
+        {
+#if NET8_0_OR_GREATER
+            ArgumentException.ThrowIfNullOrWhiteSpace(argument, paramName);
+#else
+            if (string.IsNullOrWhiteSpace(argument))
+                Throw(argument, paramName);
+
+            [DoesNotReturn]
+            static void Throw(string? argument, string? paramName)
+            {
+                ArgumentNullException.ThrowIfNull(argument, paramName);
+                throw new ArgumentException(Resources.Argument_EmptyOrWhiteSpaceString, paramName);
+            }
+#endif
+        }
+    }
+
+    /// <summary>
     /// Provides polyfill extension methods for <see cref="ArgumentNullException"/> class.
     /// </summary>
     extension(ArgumentNullException)
@@ -39,13 +104,13 @@ public static class ExceptionPolyfills
         /// If you omit this parameter, the name of <paramref name="argument"/> is used.
         /// </param>
         /// <exception cref="ArgumentNullException"><paramref name="argument"/> is <see langword="null"/>.</exception>
-#if TFF_EXCEPTION_THROWIF
+#if NET6_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIfNull([NotNull] object? argument, [CallerArgumentExpression(nameof(argument))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET6_0_OR_GREATER
             ArgumentNullException.ThrowIfNull(argument, paramName);
 #else
             if (argument is null)
@@ -59,14 +124,14 @@ public static class ExceptionPolyfills
         /// <param name="argument">The pointer argument to validate as non-null</param>
         /// <param name="paramName">The name of the parameter with which <paramref name="argument"/> corresponds.</param>
         /// <exception cref="ArgumentNullException"><paramref name="argument"/> is <see langword="null"/>.</exception>
-#if TFF_EXCEPTION_THROWIF
+#if NET7_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         [CLSCompliant(false)]
         public static unsafe void ThrowIfNull([NotNull] void* argument, [CallerArgumentExpression(nameof(argument))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET7_0_OR_GREATER
             ArgumentNullException.ThrowIfNull(argument, paramName);
 #else
             if (argument is null)
@@ -88,13 +153,13 @@ public static class ExceptionPolyfills
         /// <param name="value">The argument to validate as non-zero.</param>
         /// <param name="paramName">The name of the parameter with which <paramref name="value"/> corresponds.</param>
         /// <exception cref="ArgumentOutOfRangeException"><paramref name="value"/> is zero.</exception>
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIfZero(int value, [CallerArgumentExpression(nameof(value))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
             ArgumentOutOfRangeException.ThrowIfZero(value, paramName);
 #else
             if (value == 0)
@@ -103,13 +168,13 @@ public static class ExceptionPolyfills
         }
 
         /// <inheritdoc cref="ThrowIfZero(int, string?)"/>
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIfZero(double value, [CallerArgumentExpression(nameof(value))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
             ArgumentOutOfRangeException.ThrowIfZero(value, paramName);
 #else
             if (value == 0)
@@ -117,7 +182,7 @@ public static class ExceptionPolyfills
 #endif
         }
 
-#if !TFF_EXCEPTION_THROWIF
+#if !NET8_0_OR_GREATER
         [DoesNotReturn]
         static void ThrowZero(object value, string? paramName) =>
             throw new ArgumentOutOfRangeException(paramName, value, string.Format(Resources.ArgumentOutOfRange_Generic_MustBeNonZero, paramName, value));
@@ -133,13 +198,13 @@ public static class ExceptionPolyfills
         /// <param name="value">The argument to validate as non-negative.</param>
         /// <param name="paramName">The name of the parameter with which <paramref name="value"/> corresponds.</param>
         /// <exception cref="ArgumentOutOfRangeException"><paramref name="value"/> is negative.</exception>
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIfNegative(int value, [CallerArgumentExpression(nameof(value))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
             ArgumentOutOfRangeException.ThrowIfNegative(value, paramName);
 #else
             if (value < 0)
@@ -148,13 +213,13 @@ public static class ExceptionPolyfills
         }
 
         /// <inheritdoc cref="ThrowIfNegative(int, string?)"/>
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIfNegative(double value, [CallerArgumentExpression(nameof(value))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
             ArgumentOutOfRangeException.ThrowIfNegative(value, paramName);
 #else
             if (value < 0)
@@ -162,7 +227,7 @@ public static class ExceptionPolyfills
 #endif
         }
 
-#if !TFF_EXCEPTION_THROWIF
+#if !NET8_0_OR_GREATER
         [DoesNotReturn]
         static void ThrowNegative(object value, string? paramName) =>
             throw new ArgumentOutOfRangeException(paramName, value, string.Format(Resources.ArgumentOutOfRange_Generic_MustBeNonNegative, paramName, value));
@@ -178,13 +243,13 @@ public static class ExceptionPolyfills
         /// <param name="value">The argument to validate as non-zero or non-negative.</param>
         /// <param name="paramName">The name of the parameter with which <paramref name="value"/> corresponds.</param>
         /// <exception cref="ArgumentOutOfRangeException"><paramref name="value"/> is negative or zero.</exception>
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIfNegativeOrZero(int value, [CallerArgumentExpression(nameof(value))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
             ArgumentOutOfRangeException.ThrowIfNegativeOrZero(value, paramName);
 #else
             if (value <= 0)
@@ -193,13 +258,13 @@ public static class ExceptionPolyfills
         }
 
         /// <inheritdoc cref="ThrowIfNegativeOrZero(int, string?)"/>
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIfNegativeOrZero(double value, [CallerArgumentExpression(nameof(value))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
             ArgumentOutOfRangeException.ThrowIfNegativeOrZero(value, paramName);
 #else
             if (value <= 0)
@@ -207,7 +272,7 @@ public static class ExceptionPolyfills
 #endif
         }
 
-#if !TFF_EXCEPTION_THROWIF
+#if !NET8_0_OR_GREATER
         [DoesNotReturn]
         static void ThrowNegativeOrZero(object value, string? paramName) =>
             throw new ArgumentOutOfRangeException(paramName, value, string.Format(Resources.ArgumentOutOfRange_Generic_MustBeNonNegativeNonZero, paramName, value));
@@ -224,13 +289,13 @@ public static class ExceptionPolyfills
         /// <param name="other">The value to compare with <paramref name="value"/>.</param>
         /// <param name="paramName">The name of the parameter with which <paramref name="value"/> corresponds.</param>
         /// <exception cref="ArgumentOutOfRangeException"><paramref name="value"/> is greater than <paramref name="other"/>.</exception>
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIfGreaterThan(int value, int other, [CallerArgumentExpression(nameof(value))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
             ArgumentOutOfRangeException.ThrowIfGreaterThan(value, other, paramName);
 #else
             if (value > other)
@@ -239,13 +304,13 @@ public static class ExceptionPolyfills
         }
 
         /// <inheritdoc cref="ThrowIfGreaterThan(int, int, string?)"/>
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIfGreaterThan(double value, double other, [CallerArgumentExpression(nameof(value))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
             ArgumentOutOfRangeException.ThrowIfGreaterThan(value, other, paramName);
 #else
             if (value > other)
@@ -253,7 +318,7 @@ public static class ExceptionPolyfills
 #endif
         }
 
-#if !TFF_EXCEPTION_THROWIF
+#if !NET8_0_OR_GREATER
         [DoesNotReturn]
         static void ThrowGreater(object value, object other, string? paramName) =>
             throw new ArgumentOutOfRangeException(paramName, value, string.Format(Resources.ArgumentOutOfRange_Generic_MustBeLessOrEqual, paramName, value, other));
@@ -270,13 +335,13 @@ public static class ExceptionPolyfills
         /// <param name="other">The value to compare with <paramref name="value"/>.</param>
         /// <param name="paramName">The name of the parameter with which <paramref name="value"/> corresponds.</param>
         /// <exception cref="ArgumentOutOfRangeException"><paramref name="value"/> is greater than or equal to <paramref name="other"/>.</exception>
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIfGreaterThanOrEqual(int value, int other, [CallerArgumentExpression(nameof(value))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
             ArgumentOutOfRangeException.ThrowIfGreaterThanOrEqual(value, other, paramName);
 #else
             if (value >= other)
@@ -285,13 +350,13 @@ public static class ExceptionPolyfills
         }
 
         /// <inheritdoc cref="ThrowIfGreaterThanOrEqual(int, int, string?)"/>
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIfGreaterThanOrEqual(double value, double other, [CallerArgumentExpression(nameof(value))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
             ArgumentOutOfRangeException.ThrowIfGreaterThanOrEqual(value, other, paramName);
 #else
             if (value >= other)
@@ -299,7 +364,7 @@ public static class ExceptionPolyfills
 #endif
         }
 
-#if !TFF_EXCEPTION_THROWIF
+#if !NET8_0_OR_GREATER
         [DoesNotReturn]
         static void ThrowGreaterEqual(object value, object other, string? paramName) =>
             throw new ArgumentOutOfRangeException(paramName, value, string.Format(Resources.ArgumentOutOfRange_Generic_MustBeLess, paramName, value, other));
@@ -316,13 +381,13 @@ public static class ExceptionPolyfills
         /// <param name="other">The value to compare with <paramref name="value"/>.</param>
         /// <param name="paramName">The name of the parameter with which <paramref name="value"/> corresponds.</param>
         /// <exception cref="ArgumentOutOfRangeException"><paramref name="value"/> is less than <paramref name="other"/>.</exception>
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIfLessThan(int value, int other, [CallerArgumentExpression(nameof(value))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
             ArgumentOutOfRangeException.ThrowIfLessThan(value, other, paramName);
 #else
             if (value < other)
@@ -331,13 +396,13 @@ public static class ExceptionPolyfills
         }
 
         /// <inheritdoc cref="ThrowIfLessThan(int, int, string?)"/>
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIfLessThan(double value, double other, [CallerArgumentExpression(nameof(value))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
             ArgumentOutOfRangeException.ThrowIfLessThan(value, other, paramName);
 #else
             if (value < other)
@@ -345,7 +410,7 @@ public static class ExceptionPolyfills
 #endif
         }
 
-#if !TFF_EXCEPTION_THROWIF
+#if !NET8_0_OR_GREATER
         [DoesNotReturn]
         static void ThrowLess(object value, object other, string? paramName) =>
             throw new ArgumentOutOfRangeException(paramName, value, string.Format(Resources.ArgumentOutOfRange_Generic_MustBeGreaterOrEqual, paramName, value, other));
@@ -362,13 +427,13 @@ public static class ExceptionPolyfills
         /// <param name="other">The value to compare with <paramref name="value"/>.</param>
         /// <param name="paramName">The name of the parameter with which <paramref name="value"/> corresponds.</param>
         /// <exception cref="ArgumentOutOfRangeException"><paramref name="value"/> is less than or equal to <paramref name="other"/>.</exception>
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIfLessThanOrEqual(int value, int other, [CallerArgumentExpression(nameof(value))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
             ArgumentOutOfRangeException.ThrowIfLessThanOrEqual(value, other, paramName);
 #else
             if (value <= other)
@@ -377,13 +442,13 @@ public static class ExceptionPolyfills
         }
 
         /// <inheritdoc cref="ThrowIfLessThanOrEqual(int, int, string?)"/>
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIfLessThanOrEqual(double value, double other, [CallerArgumentExpression(nameof(value))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
             ArgumentOutOfRangeException.ThrowIfLessThanOrEqual(value, other, paramName);
 #else
             if (value <= other)
@@ -391,7 +456,7 @@ public static class ExceptionPolyfills
 #endif
         }
 
-#if !TFF_EXCEPTION_THROWIF
+#if !NET8_0_OR_GREATER
         [DoesNotReturn]
         static void ThrowLessEqual(object value, object other, string? paramName) =>
             throw new ArgumentOutOfRangeException(paramName, value, string.Format(Resources.ArgumentOutOfRange_Generic_MustBeGreater, paramName, value, other));
@@ -408,13 +473,13 @@ public static class ExceptionPolyfills
         /// <param name="other">The value to compare with <paramref name="value"/>.</param>
         /// <param name="paramName">The name of the parameter with which <paramref name="value"/> corresponds.</param>
         /// <exception cref="ArgumentOutOfRangeException"><paramref name="value"/> is equal to <paramref name="other"/>.</exception>
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIfEqual(int value, int other, [CallerArgumentExpression(nameof(value))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
             ArgumentOutOfRangeException.ThrowIfEqual(value, other, paramName);
 #else
             if (value == other)
@@ -423,13 +488,13 @@ public static class ExceptionPolyfills
         }
 
         /// <inheritdoc cref="ThrowIfEqual(int, int, string?)"/>
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIfEqual(double value, double other, [CallerArgumentExpression(nameof(value))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
             ArgumentOutOfRangeException.ThrowIfEqual(value, other, paramName);
 #else
             if (value == other)
@@ -437,7 +502,7 @@ public static class ExceptionPolyfills
 #endif
         }
 
-#if !TFF_EXCEPTION_THROWIF
+#if !NET8_0_OR_GREATER
         [DoesNotReturn]
         static void ThrowEqual(object? value, object? other, string? paramName) =>
             throw new ArgumentOutOfRangeException(paramName, value, string.Format(Resources.ArgumentOutOfRange_Generic_MustBeNotEqual, paramName, value ?? "null", other ?? "null"));
@@ -454,13 +519,13 @@ public static class ExceptionPolyfills
         /// <param name="other">The value to compare with <paramref name="value"/>.</param>
         /// <param name="paramName">The name of the parameter with which <paramref name="value"/> corresponds.</param>
         /// <exception cref="ArgumentOutOfRangeException"><paramref name="value"/> is equal to <paramref name="other"/>.</exception>
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIfNotEqual(int value, int other, [CallerArgumentExpression(nameof(value))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
             ArgumentOutOfRangeException.ThrowIfNotEqual(value, other, paramName);
 #else
             if (value != other)
@@ -469,13 +534,13 @@ public static class ExceptionPolyfills
         }
 
         /// <inheritdoc cref="ThrowIfNotEqual(int, int, string?)"/>
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIfNotEqual(double value, double other, [CallerArgumentExpression(nameof(value))] string? paramName = null)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET8_0_OR_GREATER
             ArgumentOutOfRangeException.ThrowIfNotEqual(value, other, paramName);
 #else
             if (value != other)
@@ -483,7 +548,7 @@ public static class ExceptionPolyfills
 #endif
         }
 
-#if !TFF_EXCEPTION_THROWIF
+#if !NET8_0_OR_GREATER
         [DoesNotReturn]
         static void ThrowNotEqual(object? value, object? other, string? paramName) =>
             throw new ArgumentOutOfRangeException(paramName, value, string.Format(Resources.ArgumentOutOfRange_Generic_MustBeEqual, paramName, value ?? "null", other ?? "null"));
@@ -503,13 +568,13 @@ public static class ExceptionPolyfills
         /// <param name="condition">The condition to evaluate.</param>
         /// <param name="instance">The object whose type's full name should be included in any resulting <see cref="ObjectDisposedException"/>.</param>
         /// <exception cref="ObjectDisposedException">The <paramref name="condition"/> is <see langword="true"/>.</exception>
-#if TFF_EXCEPTION_THROWIF
+#if NET7_0_OR_GREATER
         [EditorBrowsable(EditorBrowsableState.Never)]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
 #endif
         public static void ThrowIf([DoesNotReturnIf(true)] bool condition, object instance)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET7_0_OR_GREATER
             ObjectDisposedException.ThrowIf(condition, instance);
 #else
             if (condition)
@@ -525,7 +590,7 @@ public static class ExceptionPolyfills
         /// <param name="type">The type whose full name should be included in any resulting <see cref="ObjectDisposedException"/>.</param>
         public static void ThrowIf([DoesNotReturnIf(true)] bool condition, Type type)
         {
-#if TFF_EXCEPTION_THROWIF
+#if NET7_0_OR_GREATER
             ObjectDisposedException.ThrowIf(condition, type);
 #else
             if (condition)